Syncing With Zenpock's Fork

+++ include/fx/egg.h
@@ -2,12 +2,17 @@
 #define __FX_EGG_H__
 
 #include "common.h"
+#include "overlays/ba/assets.h"
+#include "fx/dlsmoke.h"
+#include "core2/1E8F430.h"
+#include "core2/1E93440.h"
+#include "core2/1EC8070.h"
 
 void _fxegg_entrypoint_1(f32*, f32);
 void _fxegg_entrypoint_2(f32*, f32);
 void _fxegg_entrypoint_3(f32*, f32);
 void _fxegg_entrypoint_4(f32*, f32);
-void _fxegg_entrypoint_5(f32*, s32*, f32);
+void _fxegg_entrypoint_5(f32*, f32*, f32);
 void _fxegg_entrypoint_6(f32*, f32, f32*);
 void _fxegg_entrypoint_7(f32*, f32*, f32);
 void _fxegg_entrypoint_8(f32*, f32*, f32);
